refactor: expose CWindow::hasPopupAt
